Development (#1)
* Added hashing function which wil be used for generating an integer from a string

* Untested: switched wchar_t* uses to std::wstring

* Added ProfileInfo::GetGroupHash method

* Switched around the if to be more readable and use the macro I created

* Removed all debug shit as I know driver calls are out the ass right now, I will readd them later

* Modified ItemTemplate to be easier to use

* Disabled optimizations in debug... why were they on? :(

* Switched from intptr_t to uint64_t

* When adding new settings to the config use TRY_READ, this will automatically return the default value of that type if it is not present in the config instead of throwing an exception

* Changed intptr_t to uint64_t, added box ESP (scale it yourself, 256 works fine for me), added group id, removed autistic way of converting wchar to mbchar, made Daniel's code readable, undid the sorting as it was making it difficult to debug

* Important: I messed up doing some of the config stuff, and I made a macro because why not I like them

* Untested: changed Memory::ReadList to return a std::vector<T> instead of an std::tuple.

* Moved the code to get the position from a transform from Player to Hacks as it is not only relevant to Player

* Added loot ESP, added some more macro helpers in the settings, switched a lot of the value macros to be const values instead, removed Global::lootESPFilters as it is part of the config, might replace the for loop for clearing the filterInputBuffer with a memset, I'm not actually sure how fast it is but regardless it looks cleaner in my mind

* fixed lag issue with loot esp large quantity of items

* Fixt retard code

* 👍

---------

+++ DriverInterop.h
@@ -9,12 +9,12 @@ public:
     DriverInterop(const DriverInterop& interop) = delete;
     ~DriverInterop();
 
-    bool ReadRaw(intptr_t address, void* pBuffer, unsigned long size);
-    bool WriteRaw(intptr_t address, void* pBuffer, unsigned long size);
+    bool ReadRaw(uint64_t address, void* pBuffer, unsigned long size);
+    bool WriteRaw(uint64_t address, void* pBuffer, unsigned long size);
 
     bool UpdateProcessId(const wchar_t* processName);
-    intptr_t GetBaseAddress();
-    intptr_t GetModuleBase();
+    uint64_t GetBaseAddress();
+    uint64_t GetModuleBase();
     bool SetTargetModule(wchar_t* moduleName);
 
 private:
