+++ InventoryController.cpp
@@ -10,11 +10,11 @@ bool InventoryController::IsInInventory() {
 Slot InventoryController::GetSlot(EquipmentSlot slot) {
 	if (!this->cachedSlots[slot].has_value()) {
 		this->cachedSlots[slot] = Slot{
-			Memory::ReadValue<intptr_t>(
+			Memory::ReadValue<uint64_t>(
 				Global::pMemoryInterface,
-				Memory::ReadChain<intptr_t>(
+				Memory::ReadChain<uint64_t>(
 					Global::pMemoryInterface, this->address, {Offsets::InventoryController::InventoryClass, Offsets::InventoryClass::EquipmentClass, 0x78}
-				) + Offsets::List::FirstEntry + sizeof(intptr_t) * slot
+				) + Offsets::List::FirstEntry + sizeof(uint64_t) * slot
 			)
 		};
 	}
