+++ Offsets.hpp
@@ -4,33 +4,33 @@
 #include <array>
 
 namespace Offsets {
-	extern intptr_t GOM;
+	extern uint64_t GOM;
 
 	namespace GameObjectManager {
-		extern intptr_t TaggedObjects;
-		extern intptr_t ActiveObjects;
+		extern uint64_t TaggedObjects;
+		extern uint64_t ActiveObjects;
 	}
 
 	namespace GameWorld {
-		extern intptr_t RegisteredPlayers;
+		extern uint64_t RegisteredPlayers;
 	}
 
 	namespace List {
-		extern intptr_t Base;
-		extern intptr_t Count;
-		extern intptr_t FirstEntry;
+		extern uint64_t Base;
+		extern uint64_t Count;
+		extern uint64_t FirstEntry;
 	}
 
 	namespace Player {
-		extern intptr_t InventoryController;
+		extern uint64_t InventoryController;
 	}
 
 	namespace InventoryController {
-		extern intptr_t InventoryClass;
-		extern intptr_t IsInInventory;
+		extern uint64_t InventoryClass;
+		extern uint64_t IsInInventory;
 	}
 
 	namespace InventoryClass {
-		extern intptr_t EquipmentClass;
+		extern uint64_t EquipmentClass;
 	}
 }
